Move code to dissemination domain
